Fixed a bug in RadioButton, unifying Icon & IconToggle into one component
