+++ lib/Avatar.js
@@ -36,24 +36,22 @@ export default class Avatar extends Component {
 
         if (icon) {
             return (
-                <View style={{ flex: 1 }}>
-                    <View style={{ width: size, height: size, borderRadius: size / 2, backgroundColor: getColor(backgroundColor), alignItems:'center', justifyContent: 'center' }}>
-                        <Icon
-                            name={icon}
-                            color={color}
-                            size={0.6 * size}
-                        />
-                    </View>
+                <View style={{ width: size, height: size, borderRadius: size / 2, backgroundColor: getColor(backgroundColor), alignItems:'center', justifyContent: 'center' }}>
+                    <Icon
+                        name={icon}
+                        color={getColor(color)}
+                        size={0.6 * size}
+                    />
                 </View>
             );
         }
 
         if (text) {
             return (
-                <View style={{ flex: 1 }}>
-                    <View style={{ width: size, height: size, borderRadius: size / 2, backgroundColor: getColor(backgroundColor), alignItems:'center', justifyContent: 'center' }}>
-                        <Text style={{ color: color }}>{text}</Text>
-                    </View>
+                <View style={{ width: size, height: size, borderRadius: size / 2, backgroundColor: getColor(backgroundColor), alignItems:'center', justifyContent: 'center' }}>
+                    <Text style={{ color: getColor(color), fontSize: size / 3 }}>
+                        {text}
+                    </Text>
                 </View>
             );
         }
