+++ lib/Checkbox.js
@@ -1,7 +1,6 @@
 import React, { Component, StyleSheet, PropTypes, Text, View, TouchableHighlight } from 'react-native';
 import { TYPO, PRIMARY, COLOR, PRIMARY_COLORS, THEME_NAME } from './config';
 import Icon from './Icon';
-import IconToggle from './IconToggle';
 
 const typos = StyleSheet.create(TYPO);
 
@@ -108,21 +107,19 @@ export default class Checkbox extends Component {
                 activeOpacity={1}
             >
                 <View style={styles.container}>
-                    <IconToggle
+                    <Icon
                         disabled={disabled}
-                        color={CURR_COLOR}
+                        pressed={{color: CURR_COLOR}}
                         onPress={this.onPress}
-                    >
-                        <Icon name={this.state.checked ? 'check-box' : 'check-box-outline-blank'}
-                              size={24}
-                              color={CURR_COLOR}
-                              key={value}
-                              style={{
-                                  opacity: OPACITY,
-                                  margin: 16,
-                              }}
-                        />
-                    </IconToggle>
+                        name={this.state.checked ? 'check-box' : 'check-box-outline-blank'}
+                        size={24}
+                        color={CURR_COLOR}
+                        key={value}
+                        style={{
+                          opacity: OPACITY,
+                          margin: 16,
+                        }}
+                    />
                     <View
                         style={styles.labelContainer}
                         onPress={this.onPress}
