+++ lib/RadioButton.js
@@ -1,6 +1,5 @@
 import React, { Component, StyleSheet, PropTypes, Text, View, TouchableHighlight } from 'react-native';
 import Icon from './Icon';
-import IconToggle from './IconToggle';
 import { TYPO, PRIMARY, COLOR, THEME_NAME, PRIMARY_COLORS } from './config';
 
 const typos = StyleSheet.create(TYPO);
@@ -22,6 +21,21 @@ export default class RadioButton extends Component {
         primary: PRIMARY,
         disabled: false
     };
+
+    constructor(props){
+        super(props);
+        this._onPress = this._onPress.bind(this);
+    }
+
+    _onPress(){
+        const {checked, disabled, value, onSelect } = this.props;
+        if(disabled && !checked){
+            return;
+        }
+        if(onSelect){
+            onSelect(value);
+        }
+    }
 
     render() {
         const { theme, primary, value, checked, disabled, onSelect } = this.props;
@@ -80,25 +94,24 @@ export default class RadioButton extends Component {
 
         return (
             <TouchableHighlight
-                onPress={() => { disabled && !checked ? null : onSelect(value) }}
+                onPress={this._onPress}
                 underlayColor={disabled ? 'rgba(0,0,0,0)' : UNDERLAY_COLOR}
                 activeOpacity={1}
             >
                 <View style={styles.container}>
-                    <IconToggle
+                    <Icon
                         disabled={disabled}
+                        pressed={{color: CURR_COLOR}}
+                        onPress={this._onPress}
+                        name={checked ? 'radio-button-checked' : 'radio-button-unchecked'}
+                        size={24}
                         color={CURR_COLOR}
-                        onPress={() => { disabled && !checked ? null : onSelect(value) }}
-                    >
-                        <Icon name={checked ? 'radio-button-checked' : 'radio-button-unchecked'}
-                              size={24}
-                              color={CURR_COLOR}
-                              style={{
-                                  opacity: OPACITY,
-                                  margin: 16
-                              }}
-                        />
-                    </IconToggle>
+                        style={{
+                          opacity: OPACITY,
+                          margin: 16
+                        }}
+                    />
+
                     <View style={styles.labelContainer}>
                         <Text
                             style={[
