+++ lib/Toolbar.js
@@ -2,7 +2,6 @@ import React, { Component, PropTypes, View, Text } from 'react-native';
 import { TYPO, PRIMARY, THEME_NAME, PRIMARY_COLORS } from './config';
 import { getColor } from './helpers';
 import Icon from './Icon';
-import IconToggle from './IconToggle';
 
 export default class Toolbar extends Component {
 
@@ -96,17 +95,15 @@ export default class Toolbar extends Component {
                         return (
                             <IconToggle
                                 key={i}
-                                color={styleMap.rightIconColor}
+                                pressed={{color: styleMap.rightIconColor}}
                                 badge={action.badge}
                                 onPress={action.onPress}
                                 disabled={action.disabled}
-                            >
-                                <Icon name={action.icon}
-                                      size={24}
-                                      color={styleMap.rightIconColor}
-                                      style={[styles.rightIcon, rightIconStyle, action.disabled ? { opacity: opacityMap[theme] } : null]}
-                                />
-                            </IconToggle>
+                                name={action.icon}
+                                size={24}
+                                color={styleMap.rightIconColor}
+                                style={[styles.rightIcon, rightIconStyle, action.disabled ? { opacity: opacityMap[theme] } : null]}
+                            />
                         );
                     })
                 }
