+++ lib/index.js
@@ -7,7 +7,6 @@ export { default as CheckboxGroup } from './CheckboxGroup';
 export { default as Drawer } from './Drawer';
 export { default as Divider } from './Divider';
 export { default as Icon } from './Icon';
-export { default as IconToggle } from './IconToggle';
 //export { default as List } from './List';
 export { default as RadioButton } from './RadioButton';
 export { default as Ripple } from './Ripple';
